#  smoothquant fp8 support