--- a/autosmoothquant/examples/test_model.py
+++ b/autosmoothquant/examples/test_model.py
@@ -3,7 +3,7 @@ import torch
 import argparse
 import json
 
-from autosmoothquant.models import Int8LlamaForCausalLM, Int8OPTForCausalLM, Int8BaichuanForCausalLM, Int8MixtralForCausalLM
+from autosmoothquant.models import QuantizedLlamaForCausalLM, Int8OPTForCausalLM, Int8BaichuanForCausalLM, Int8MixtralForCausalLM
 from autosmoothquant.utils import parse_quant_config
 from transformers import AutoTokenizer
 
@@ -16,7 +16,7 @@ def parse_args():
     parser.add_argument('--model-class', type=str,
                         default='llama', help='currently support: llama, baichuan, opt, mixtral')
     parser.add_argument('--prompt', type=str,
-                        default='You are right, But Genshin Impact is', help='prompts')   
+                        default='You are right, But Genshin Impact is', help='prompts')
     args = parser.parse_args()
     return args
 
@@ -32,7 +32,7 @@ def main():
     # Consider setting the default data type to torch.float16 to speed up, but this may decrease model performance.
     # torch.set_default_dtype(torch.float16)
     if args.model_class == "llama":
-      model = Int8LlamaForCausalLM.from_pretrained(args.model_path, quant_config, attn_implementation="eager", device_map="sequential")
+      model = QuantizedLlamaForCausalLM.from_pretrained(args.model_path, quant_config, attn_implementation="eager", device_map="sequential")
     elif args.model_class == "baichuan":
       model = Int8BaichuanForCausalLM.from_pretrained(args.model_path, quant_config, attn_implementation="eager", device_map="sequential")
     elif args.model_class == "opt":