--- a/autosmoothquant/utils/__init__.py
+++ b/autosmoothquant/utils/__init__.py
@@ -1,2 +1,3 @@
 from .datautils import get_loaders
-from .utils import get_config, get_model_architecture, build_model_and_tokenizer, parse_quant_config
+from .utils import get_config, get_model_architecture, build_model_and_tokenizer, parse_quant_config, setup_seed
+from .eval_utils import pattern_match, update_results