--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,8 @@
-transformers == 4.36.2
+transformers == 4.42.3
+torch>=2.2
 datasets
 accelerate
 icecream
+tqdm
+zstandard
+lm_eval