hi~ @AniZpZ 

We have implemented smoothquant fp8 quantization. Compared with the int8 model, the fp8 model has significantly better model performance. At the same time, in comparison with the AutoFP8 method, smoothquant fp8 has a slightly lower perplexity in WikiText2. On some zero-shot tasks, smoothquant fp8 has better performance.

### Model Performance
We evaluated the model performance on WikiText2 and five zero-shot tasks. Currently only for llama2-7b.
![image](https://github.com/user-attachments/assets/9d107b25-8005-4b57-9b5b-a47a2c131d65)

![image](https://github.com/user-attachments/assets/b942a2e3-124f-4d6e-9ad3-83c00d8aca1d)

### TODO list

- support for more models
- more model evaluations
- Automatically search for smooth strength param