@huangtingwei9988 2025-04-01
Benchmark Result
```

# H20 tp=1 
python -m sglang.bench_one_batch_server --model None --base-url http://localhost:8188 --batch-size 64 --input-len 1024 --output-len 1024

# Before
batch size: 16 latency: 1.86 s  output throughput: 137.61 token/s  (input + output) throughput: 8944.92 token/s
batch size: 64 latency: 26.59s  output throughput: 2464.98 token/s  (input + output) throughput: 4929.96 token/s

# After
batch size: 16 latency: 1.87 s  output throughput: 136.53 token/s  (input + output) throughput: 8874.69 token/s
batch size: 64 latency: 26.57 s  output throughput: 2466.79 token/s  (input + output) throughput: 4933.58 token/s

```