@huangtingwei9988 2025-04-05
## Motivation

When a hash calculation is performed on a tensor, it is usually calculated based on the tensor's memory address. When two tensors have the same value but different memory addresses, their hash values ​​are different.
```

import torch

randn_tensor = torch.randn(4, 1024)
print(hash(randn_tensor))
print(hash(randn_tensor.clone()))

# result 
# 139792403318000
# 139786750322192

```
For sglang's hash_feature [https://github.com/sgl-project/sglang/blob/ca8d02abd520194bc0d182d62aeb39d9496ca91d/python/sglang/srt/managers/schedule_batch.py#L204](url)
The same multimodal input will get different hash values
```

import torch


tensor_list_1 = []
for i in range(10):
    tensor_list_1.append(torch.randn(4, 1024))
print(hash(tuple(tensor_list_1)))

tensor_list_2 = []
for i in range(10):
    tensor_list_2.append(tensor_list_1[i].clone())
print(hash(tuple(tensor_list_2)))


# result 
# -2006800841837242759
# 7092966304206677122
```


## Modifications

- [ y] fix bug
- [ ]   better performance